requirements.txt
astroid==2.3.3
autopep8==0.8
backcall==0.1.0
click==7.1.1
decorator==4.4.2
Flask==1.1.1
Flask-JWT-Extended==3.24.1
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.6.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
python-dotenv==0.12.0
redis==3.4.1
six==1.14.0
traitlets==4.3.3
typed-ast==1.4.1
uWSGI==2.0.18
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.11.2
